[HADOOP-7814] Publish counters via dr-elephant
The java file DaliMetricsAPI.java has a flavor of the APIs that we will be exposing from the dali library.
We can split these classes into individual files when we move this functionality to the dali library.

Changed start script to look for a config file that configures a publisher. If the file is present,
then dr-elephant is started with an option that has the file name. If the file is not present,
then the behavior is unchanged (i.e. no metrics are published).

If the file is parsed correctly then dr-elephant publishes metrics in HDFS (one avro file per job)
for jobs that are configured to publish the metrics.

The job needs to set something like mapreduce.job.publish-counters='org.apache.hadoop.examples.WordCount$AppCounter:*'
to publish all counters in the group given. The format is : 'groupName:counterName' where counterName can be an
asterisk to indicate all counters in the group. See the class DaliMetricsAPI.CountersToPublish

The HDFSPublisher is configured with a base path under which metrics are published. The date/hour hierarchy is added
to the base path.

The XML file for configuring dr-elephant is checked in as a template. A config file needs to be added to the
'conf' path of dr-elephant (manually, as per meeting with hadoop-admin) on clusters where we want dr-elephant
to publish metrics.

RB=409443
BUGS=HADOOP-7814
R=fli,csteinba,mwagner,cbotev,ahsu
A=fli,ahsu