Fixed the build script.
Changed the script so it prints out what it is doing.

I suspect that somehow the directory name was not getting set correctly. Used a
one-liner that seems to work on both Mac and Linux to determine directory of a script.

Used braces to delimit the shell variable names. May be random variables were getting set?

The zip command included a -x that pretty much amounted to nothing, since it excluded '*'.
Changed the zip command to be more specific.

Finally, added a trap command that fields ^C and exits the script.

It seems to survive ctrl-C now.

RB=399693
BUGS=HADOOP-7816
R=fli,akrai
A=fli