[HADOOP-8640] Fix hang in dr-elephant due to h1 and h2 incompatibility
hadoop-1 does not have JobStatus.getFinishTime(). This causes dr-elephant to hang.

Set the start time to be same as finish time for h1 jobs.

For consistency, reverted to the old method of scraping the job tracker url so that we get only
start time, and set the finish time to be equal to start time for retired jobs as well.

RB=417975
BUGS=HADOOP-8640
R=fli,mwagner
A=fli